extract `confirmdestroy` to external file
