+++ src/commands/destroy.js
@@ -8,6 +8,7 @@ const { destroyBuckets } = require('../gcp/destroyBuckets');
 const configureCredentials = require('../utils/configureCredentials');
 const { functionExists } = require('../gcp/validateIntegrationName');
 const ora = require('ora');
+const confirmDestroy = require('../utils/confirmDestroy');
 
 module.exports = async (args) => {
   await confirmDestroy();
@@ -76,41 +77,4 @@ const getGCEInstance = async (integrationName) => {
     filter: `name eq ${integrationName}`,
   });
   return instances[0];
-}
-
-const readline = require('readline-sync');
-
-const confirmDestroy = async () => {
-  const confirmationMessage = ` You are about to delete these resources:
-  * Remove the integration from the list of mashr integrations
-  * Destroy all related GCP resources, including:
-    - The staging and archive GCS buckets
-    - The Cloud Function
-    - The GCE instance
-Since the archive and staging GCS buckets will be permanently destroyed,
-Consider moving your data before running this command.
-Continue (y/n)> `
-
-  let response = readline.question(confirmationMessage);
-  response = response.trim();
-
-  if (response === 'y') { response = 'yes'}
-  if (response === 'n') { response = 'no'}
-
-  switch (response) {
-    case 'yes':
-      // console.log('goodbye world!');
-      // process.exit(0);
-      break;
-    case 'no':
-      // console.log('hello world!');
-      process.exit(0);
-      break;
-    default:
-      console.log('');
-      console.log(`Please choose 'y' or 'n'.`);
-      console.log('');
-      confirmDestroy();
-      break;
-  }
 }
